♻️ refactor : gathering 관련 타입분리
