+++ components/common/GatheringList.tsx
@@ -1,23 +1,5 @@
 import GatheringItem from "@components/common/GatheringItem";
-
-export interface Gathering {
-  teamId: number;
-  id: number;
-  type: string;
-  name: string;
-  dateTime: string;
-  registrationEnd: string;
-  location: string;
-  participantCount: number;
-  capacity: number;
-  image: string;
-  createdBy: number | null;
-  canceledAt: string | null;
-}
-
-export interface GatheringListProps {
-  gatherings: Gathering[];
-}
+import { GatheringListProps } from "@customTypes/gathering";
 
 export default function GatheringList({ gatherings }: GatheringListProps) {
   return (
